+++ CHANGELOG.md
@@ -8,6 +8,8 @@ _unreleased_
   - Header with search sticks to top of page if scrolled down
   - Add entry form: Align style with application
   - Improve various UI parts for mobile phone displays
+- Other changes
+  - Translate UI to german
 
 
 ## 0.0.6